+++ subsys/net/conn_mgr/conn_mgr_connectivity.c
@@ -9,28 +9,7 @@ LOG_MODULE_REGISTER(conn_mgr_conn, CONFIG_NET_CONNECTION_MANAGER_LOG_LEVEL);
 
 #include <zephyr/net/net_if.h>
 #include <zephyr/net/conn_mgr_connectivity.h>
-
-/**
- * @brief Retrieves the conn_mgr binding struct for a provided iface if it exists.
- *
- * Bindings for connectivity implementations with missing API structs are ignored.
- *
- * @param iface - bound network interface to obtain the binding struct for.
- * @return struct conn_mgr_conn_binding* Pointer to the retrieved binding struct if it exists,
- *	   NULL otherwise.
- */
-static inline struct conn_mgr_conn_binding *conn_mgr_if_get_binding(struct net_if *iface)
-{
-	STRUCT_SECTION_FOREACH(conn_mgr_conn_binding, binding) {
-		if (iface == binding->iface) {
-			if (binding->impl->api) {
-				return binding;
-			}
-			return NULL;
-		}
-	}
-	return NULL;
-}
+#include "conn_mgr_private.h"
 
 int conn_mgr_if_connect(struct net_if *iface)
 {
